refactor: separated datamodels for ontologies to separate library
