+++ apps/dispatch-service/src/app/results/log.service.ts
@@ -2,8 +2,7 @@ import { Injectable, Logger } from '@nestjs/common';
 
 import YAML from 'yaml';
 import { SimulationRunService } from '@biosimulations/api-nest-client';
-import { CombineArchiveLog } from '@biosimulations/datamodel/api';
-import { SimulationRunLogStatus } from '@biosimulations/datamodel/common';
+import { CombineArchiveLog, SimulationRunLogStatus } from '@biosimulations/datamodel/common';
 import { SshService } from '../services/ssh/ssh.service';
 @Injectable()
 export class LogService {
