+++ apps/dispatch-service/tsconfig.spec.json
@@ -3,8 +3,7 @@
   "compilerOptions": {
     "outDir": "../../dist/out-tsc",
     "module": "commonjs",
-    "types": ["jest", "node"],
-    "resolveJsonModule": true
+    "types": ["jest", "node"]
   },
   "include": ["**/*.spec.ts", "**/*.d.ts"]
 }
