+++ apps/mail-service/tsconfig.app.json
@@ -4,8 +4,7 @@
     "outDir": "../../dist/out-tsc",
     "types": ["node"],
     "emitDecoratorMetadata": true,
-    "target": "es2015",
-    "resolveJsonModule": true
+    "target": "es2015"
   },
   "exclude": ["**/*.spec.ts"],
   "include": ["**/*.ts"]
