+++ libs/datamodel/api/src/lib/common/index.ts
@@ -4,9 +4,7 @@ export * from './taxon';
 export * from './ontology.dto';
 export * from './person.dto';
 export * from './externalreferences.dto';
-export * from './ontologyId.dto';
 export * from './url';
 export * from './dependentPackage';
-export * from './funding';
 export * from './archiveMetadata';
 export * from './commonDefinitions';
