+++ libs/datamodel/api/src/lib/core/index.ts
@@ -4,5 +4,3 @@ export * from './specifications';
 export * from './project';
 export * from './simulationRun';
 export * from './simulationRunReport';
-export * from './simulationRunLog';
-export * from './simulator';
