+++ libs/datamodel/api/src/lib/simulators/index.ts
@@ -1,6 +1,3 @@
-export * from './algorithm';
-export * from './algorithmParameter';
 export * from './biosimulatorsMeta';
 export * from './cli';
 export * from './pythonApi';
-export * from './image';
