+++ libs/datamodel/common/src/lib/core/simulationRunLog.ts
@@ -73,3 +73,8 @@ export interface CombineArchiveLog {
   duration: number | null;
   sedDocuments: SedDocumentLog[] | null;
 }
+
+export class CreateSimulationRunLogBody {
+  simId!: string;
+  log!: CombineArchiveLog;
+}
