+++ libs/ontology/datamodel/.eslintrc
@@ -1,0 +1,7 @@
+{
+    "extends": "../../../.eslintrc.json",
+    "rules": {},
+    "ignorePatterns": [
+        "!**/*"
+    ]
+}
