+++ libs/ontology/datamodel/README.md
@@ -1,0 +1,7 @@
+# ontology-datamodel
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `ng test ontology-datamodel` to execute the unit tests via [Jest](https://jestjs.io).
