+++ libs/ontology/datamodel/jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+  preset: '../../../jest.preset.js',
+  globals: {
+    'ts-jest': { tsconfig: '<rootDir>/tsconfig.spec.json' },
+  },
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
+  coverageDirectory: '../../../coverage/libs/ontology/datamodel',
+  displayName: 'ontology-datamodel',
+};
