+++ libs/ontology/datamodel/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/ontology-datamodel';
