+++ libs/ontology/datamodel/src/lib/common/index.ts
@@ -1,0 +1,2 @@
+export * from './ontology.dto';
+export * from './ontologyId.dto';
