+++ libs/ontology/datamodel/src/lib/logs/index.ts
@@ -1,0 +1,1 @@
+export * from './simulationRunLog';
