+++ libs/ontology/datamodel/src/lib/ontology-datamodel.ts
@@ -1,0 +1,3 @@
+export * from './common';
+export * from './simulators';
+export * from './logs';
