+++ libs/ontology/datamodel/src/lib/simulators/image.ts
