+++ libs/ontology/datamodel/tsconfig.lib.json
@@ -1,10 +1,8 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "types": ["node"],
-    "emitDecoratorMetadata": true,
-    "target": "es2015"
+    "outDir": "../../../dist/out-tsc",
+    "types": []
   },
   "exclude": ["**/*.spec.ts"],
   "include": ["**/*.ts"]
