+++ libs/ontology/datamodel/tsconfig.spec.json
